Add files via upload
In progress still.

This is a command that is to be run in Task Scheduler to discover when the reverse shell was executed. I have not found a way to identify powercat yet. The command is to obvious and may be changed. It seems like that may be the only way to identify it